Move component into feature
