--- packages/client/src/components/accordion-task/index.tsx
@@ -1,2 +1,0 @@
-export { default } from './accordion-task';
-export * from './accordion-task';
