+++ packages/client/src/feature/task/index.ts
@@ -1,0 +1,2 @@
+export { default } from './task-accordion';
+export * as TaskAccordion from './task-accordion';
