+++ packages/client/src/feature/task/task-accordion/index.ts
@@ -1,0 +1,2 @@
+export { default } from './task-accordion';
+export * from './task-accordion';
